refactor: move BottomNavWrapper to _components and add ContentWrapper for login layout
